src/backend/requirements.txt
# FastAPI 프레임워크
fastapi==0.115.5
uvicorn[standard]==0.34.0

# 데이터베이스
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Pydantic 설정
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Redis & Celery (워커 연동)
redis==5.2.1
celery==5.4.0

# HTTP 클라이언트
httpx==0.28.1
aiohttp==3.11.10

# JWT 인증
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# 암호화
cryptography==42.0.0

# 유틸리티
python-dotenv==1.0.1
